feat(ticket): implement basic repository, model, service, controller, and exception handling
- Created Ticket model with necessary fields.
- Implemented TicketRepository for database interaction.
- Developed TicketService to handle business logic.
- Added TicketController to manage API requests.
- Introduced custom TicketNotFoundException for error handling.

+++ raj/src/main/java/iit/lk/raj/model/Ticket.java
@@ -13,7 +13,7 @@ public class Ticket {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int ticketId;
+    private long ticketId;
 
     @Enumerated(EnumType.STRING) // Enum to store ticket types (VIP, Normal, etc.)
     private TicketType ticketType;
@@ -44,7 +44,7 @@ public class Ticket {
     }
 
     // Constructor for full initialization
-    public Ticket(int ticketId, TicketType ticketType, double ticketPrice, int ticketTotal, boolean ticketStatus) {
+    public Ticket(long ticketId, TicketType ticketType, double ticketPrice, int ticketTotal, boolean ticketStatus) {
         this.ticketId = ticketId;
         this.ticketType = ticketType;
         this.ticketPrice = ticketPrice;
